connect-api-be-into-be

+++ src/app/pages/authentication/side-login/side-login.component.ts
@@ -1,19 +1,38 @@
-import { Component } from '@angular/core';
+import { Component, Inject, inject } from '@angular/core';
 import { CoreService } from 'src/app/services/core.service';
 import { FormGroup, FormControl, Validators, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { Router, RouterModule } from '@angular/router';
 import { MaterialModule } from '../../../material.module';
+import { SideLoginService } from './side-login.service';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatButtonModule } from '@angular/material/button';
+import { MatInputModule } from '@angular/material/input';
+import { MAT_SNACK_BAR_DATA, MatSnackBar } from '@angular/material/snack-bar';
+import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
+import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'app-side-login',
   standalone: true,
-  imports: [RouterModule, MaterialModule, FormsModule, ReactiveFormsModule],
+  imports: [RouterModule, MaterialModule, FormsModule, ReactiveFormsModule, MatFormFieldModule, MatInputModule, MatButtonModule, MatProgressSpinnerModule, CommonModule],
   templateUrl: './side-login.component.html',
+  styles: [`
+    .custom-snackbar {
+      background-color: #4caf50; /* Màu nền (xanh lá) */
+      color: white; /* Màu chữ */
+    }
+  `]
 })
 export class AppSideLoginComponent {
+  private _snackBar = inject(MatSnackBar);
   options = this.settings.getOptions();
+  public loadingSpinner = false;
 
-  constructor(private settings: CoreService, private router: Router) { }
+  constructor(
+    private settings: CoreService, 
+    private router: Router,
+    private services: SideLoginService
+  ) { }
 
   form = new FormGroup({
     uname: new FormControl('', [Validators.required, Validators.minLength(5)]),
@@ -23,17 +42,95 @@ export class AppSideLoginComponent {
   get f() {
     return this.form.controls;
   }
+  ngOnInit(): void {
+  
+  }
 
-  submit() {
-    localStorage.clear();
-    // console.log(this.form.value);
-    if(this.form.value.uname === 'admin' && this.form.value.password == '123') {
-      localStorage.setItem('tokens', '1');
-      this.router.navigate(['/dashboard']);
-    }else{
-      alert('Sai thông tin đăng nhập');
-      this.router.navigate(['/dashboard']);
-    }
+  async submit() {
+   
+    // const result = await this.services.getLogin(this.form.value).subscribe((res) => {
+    //   console.log('res', res);
+    // });
+    // console.log(result);
+    this.loadingSpinner = true;
+    this.services.getLogin(this.form.value).subscribe(
+      response => {
+        if(response) {
+          
+          localStorage.clear();
+          localStorage.setItem('tokens', response.accessToken);
+          this.openSnackBar('Login successful', 'success');
+          this.loadingSpinner = false;
+          this.router.navigate(['/dashboard']);
+        }
+        
+      },
+      error => {
+        this.loadingSpinner = false;
+        this.openSnackBar('Login failed', 'error');
+        
+        
+      }
+    );
+    // localStorage.clear();
+    
+    // if(this.form.value.uname === 'admin' && this.form.value.password == '123') {
+    //   localStorage.setItem('tokens', '1');
+    //   this.router.navigate(['/dashboard']);
+    // }else{
+    //   alert('Sai thông tin đăng nhập');
+    //   this.router.navigate(['/dashboard']);
+    // }
     
   }
+
+  durationInSeconds = 3;
+  openSnackBar(data:any, status: string) {
+
+    this._snackBar.openFromComponent(SnackbarComponent, {
+      duration: this.durationInSeconds * 1000,
+      data: { message: data, status: status  }, // Truyền dữ liệu
+    });
+    
+    // this._snackBar.openFromComponent(SuccessSnackbarComponent, {
+    //   duration: this.durationInSeconds * 1000,
+    // });
+
+    // this._snackBar.openFromComponent(ErrorSnackbarComponent, {
+    //   duration: this.durationInSeconds * 1000,
+    // });
+  }
+
+  // openSnackBar(message: string, action: string) {
+  //   // this._snackBar.open(message, action);
+  //   this._snackBar.open(message, action, {
+  //     duration: 3000,
+  //     panelClass: ['custom-snackbar'] // Thêm class CSS tùy chỉnh
+  //   });
+  // }
 }
+
+
+
+
+@Component({
+  selector: 'sussess-snackbar',
+  template: `
+   <span [class]="data.status">{{ data.message }}</span>
+
+  `,
+  styles: `
+    .success {
+      color: #13deb9 !important;
+    }
+    .error {
+      color: red !important;
+    }
+  `,
+  standalone: true,
+})
+export class SnackbarComponent {
+  constructor(@Inject(MAT_SNACK_BAR_DATA) public data: any) {}
+}
+
+
